moved opportunity files
