+++ src/domain/collaboration/activity/ActivityLog/views/ActivityIcon.tsx
@@ -10,7 +10,7 @@ import {
   PersonOutlined,
 } from '@mui/icons-material';
 import { SubspaceIcon } from '@/domain/journey/subspace/icon/SubspaceIcon';
-import { OpportunityIcon } from '@/domain/journey/opportunity/icon/OpportunityIcon';
+import { OpportunityIcon } from '@/domain/space/icons/OpportunityIcon';
 import calloutIcons from '@/domain/collaboration/callout/utils/calloutIcons';
 import { WhiteboardIcon } from '@/domain/collaboration/whiteboard/icon/WhiteboardIcon';
