--- src/domain/journey/opportunity/utils/useOpportunityMetricsItems.ts
@@ -1,12 +1,0 @@
-import { MetricType } from '@/domain/platform/metrics/MetricType';
-import { MetricsItemSpec } from '@/domain/platform/metrics/utils/useMetricsItems';
-
-const OpportunityMetrics: MetricsItemSpec[] = [
-  {
-    label: 'common.members',
-    type: MetricType.Member,
-    color: 'neutralMedium',
-  },
-];
-
-export default OpportunityMetrics;
