+++ src/domain/shared/components/DashboardSections/DashboardAddButton.tsx
@@ -7,7 +7,7 @@ import { AddCircleOutlineOutlined } from '@mui/icons-material';
 import { ButtonBase, SxProps, Theme } from '@mui/material';
 import { MouseEventHandler } from 'react';
 import { useTranslation } from 'react-i18next';
-import { getIndentStyle } from '../../../journey/dashboardNavigation/utils';
+import { getIndentStyle } from '../../../space/components/dashboardNavigation/utils';
 
 interface DashboardAddButtonProps {
   url?: string;
