+++ src/domain/shared/components/SpaceIcon/SpaceIcon.tsx
@@ -1,6 +1,6 @@
 import { SpaceIcon } from '@/domain/journey/space/icon/SpaceIcon';
 import { SubspaceIcon } from '@/domain/journey/subspace/icon/SubspaceIcon';
-import { OpportunityIcon } from '@/domain/journey/opportunity/icon/OpportunityIcon';
+import { OpportunityIcon } from '@/domain/space/icons/OpportunityIcon';
 import { SpaceLevel } from '@/core/apollo/generated/graphql-schema';
 import { ComponentType } from 'react';
 import { SvgIconProps } from '@mui/material';
