+++ src/domain/space/admin/storage/useStorageAdminTree.tsx
@@ -18,7 +18,7 @@ import ImageIcon from '@mui/icons-material/Image';
 import { getProfileIcon } from '@/domain/shared/icons/profileIcons';
 import { SpaceIcon } from '@/domain/journey/space/icon/SpaceIcon';
 import { SubspaceIcon } from '@/domain/journey/subspace/icon/SubspaceIcon';
-import { OpportunityIcon } from '@/domain/journey/opportunity/icon/OpportunityIcon';
+import { OpportunityIcon } from '@/domain/space/icons/OpportunityIcon';
 import { FolderCopyOutlined } from '@mui/icons-material';
 
 export interface StorageAdminTreeItem {
