+++ src/domain/space/components/dashboardNavigation/utils.ts
@@ -1,4 +1,4 @@
-import { DashboardNavigationItem } from '../space/spaceDashboardNavigation/useSpaceDashboardNavigation';
+import { DashboardNavigationItem } from '../../../journey/space/spaceDashboardNavigation/useSpaceDashboardNavigation';
 import { gutters } from '@/core/ui/grid/utils';
 
 export const findCurrentPath = (
