+++ src/domain/space/icons/OpportunityIcon.tsx
