+++ src/domain/space/layout/tabbedLayout/Tabs/SpaceDashboard/SpaceDashboardView.tsx
@@ -12,7 +12,7 @@ import { UseCalloutsSetProvided } from '@/domain/collaboration/calloutsSet/useCa
 import useDirectMessageDialog from '@/domain/communication/messaging/DirectMessaging/useDirectMessageDialog';
 import ApplicationButton from '@/domain/community/application/applicationButton/ApplicationButton';
 import JourneyDashboardWelcomeBlock from '@/domain/journey/common/journeyDashboardWelcomeBlock/JourneyDashboardWelcomeBlock';
-import DashboardNavigation from '@/domain/journey/dashboardNavigation/DashboardNavigation';
+import DashboardNavigation from '@/domain/space/components/dashboardNavigation/DashboardNavigation';
 import { getSpaceWelcomeCache, removeSpaceWelcomeCache } from '@/domain/journey/space/createSpace/utils';
 import DashboardCalendarSection from '@/domain/shared/components/DashboardSections/DashboardCalendarSection';
 import DashboardUpdatesSection from '@/domain/shared/components/DashboardSections/DashboardUpdatesSection';
