+++ src/domain/space/pages/L2/OpportunityAboutPage.tsx
